requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
python-dotenv==1.0.1
greenlet==3.0.3
qrcode[pil]==7.4.2
Pillow==10.3.0
